Added logic to QuestionBlockObject in terms of command.
Temporary states, UsedObject is calling MarioFactory (super temporary) until a BlockFactory is made. Broken code is commented out for now.

I actually don't know how a type can switch out its type itself so I decided to make it so the QuestionBlock makes a UsedBlock and uses states to determine which block to draw at some moment in time.

If there is a better way to do it, we can implement such method later. For now, this may potentially work.