Added a few more block sprites and block objects to the block factory. The factory is now complete as far as the list of objects goes, it just can't currently instantiate all of them since they don't have constructors.
Changed IGameObject to inherit from IDraw since every game object was inheriting IDraw anyways. Then, since there was now a unified interface for game objects and drawing, the main game only has one collection of game objects, which can both update and draw.

Finally, I modified the Update method in IGameObject to return a boolean instead. I then went and made all the game objects return false except for the BreakableBrick, which is the only object currently that actually gets destroyed. For that, I added some simple logic to decide if it should return true. If an update returns true, it is added to a list of objects   which are then removed from the list of game objects at the end of the main game's udpate.